CFY-6290 Split resources_v2 module into multiple submodules
