+++ rest-service/manager_rest/rest/resources_v2/blueprints.py
@@ -1,0 +1,140 @@
+#########
+# Copyright (c) 2017 GigaSpaces Technologies Ltd. All rights reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+#  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  * See the License for the specific language governing permissions and
+#  * limitations under the License.
+#
+
+from flask_restful_swagger import swagger
+
+from manager_rest.rest import (
+    resources_v1,
+    rest_decorators,
+    rest_utils,
+)
+from manager_rest.storage import (
+    get_storage_manager,
+    models,
+)
+from manager_rest.utils import create_filter_params_list_description
+
+
+class Blueprints(resources_v1.Blueprints):
+    @swagger.operation(
+        responseClass='List[{0}]'.format(models.Blueprint.__name__),
+        nickname="list",
+        notes='Returns a list of submitted blueprints for the optionally '
+              'provided filter parameters {0}'
+        .format(models.Blueprint),
+        parameters=create_filter_params_list_description(
+            models.Blueprint.response_fields,
+            'blueprints'
+        )
+    )
+    @rest_decorators.exceptions_handled
+    @rest_decorators.marshal_with(models.Blueprint)
+    @rest_decorators.create_filters(models.Blueprint)
+    @rest_decorators.paginate
+    @rest_decorators.sortable(models.Blueprint)
+    @rest_decorators.all_tenants
+    def get(self, _include=None, filters=None, pagination=None, sort=None,
+            all_tenants=None, **kwargs):
+        """
+        List uploaded blueprints
+        """
+        return get_storage_manager().list(
+            models.Blueprint,
+            include=_include,
+            filters=filters,
+            pagination=pagination,
+            sort=sort,
+            all_tenants=all_tenants
+        )
+
+
+class BlueprintsId(resources_v1.BlueprintsId):
+
+    @swagger.operation(
+        responseClass=models.Blueprint,
+        nickname="getById",
+        notes="Returns a blueprint by its id."
+    )
+    @rest_decorators.exceptions_handled
+    @rest_decorators.marshal_with(models.Blueprint)
+    def get(self, blueprint_id, _include=None, **kwargs):
+        """
+        Get blueprint by id
+        """
+        with rest_utils.skip_nested_marshalling():
+            return super(BlueprintsId, self).get(blueprint_id=blueprint_id,
+                                                 _include=_include,
+                                                 **kwargs)
+
+    @swagger.operation(
+        responseClass=models.Blueprint,
+        nickname="upload",
+        notes="Submitted blueprint should be an archive "
+              "containing the directory which contains the blueprint. "
+              "Archive format may be zip, tar, tar.gz or tar.bz2."
+              " Blueprint archive may be submitted via either URL or by "
+              "direct upload.",
+        parameters=[{'name': 'application_file_name',
+                     'description': 'File name of yaml '
+                                    'containing the "main" blueprint.',
+                     'required': False,
+                     'allowMultiple': False,
+                     'dataType': 'string',
+                     'paramType': 'query',
+                     'defaultValue': 'blueprint.yaml'},
+                    {'name': 'blueprint_archive_url',
+                     'description': 'url of a blueprint archive file',
+                     'required': False,
+                     'allowMultiple': False,
+                     'dataType': 'string',
+                     'paramType': 'query'},
+                    {
+                        'name': 'body',
+                        'description': 'Binary form of the tar '
+                                       'gzipped blueprint directory',
+                        'required': True,
+                        'allowMultiple': False,
+                        'dataType': 'binary',
+                        'paramType': 'body'}],
+        consumes=[
+            "application/octet-stream"
+        ]
+
+    )
+    @rest_decorators.exceptions_handled
+    @rest_decorators.marshal_with(models.Blueprint)
+    def put(self, blueprint_id, **kwargs):
+        """
+        Upload a blueprint (id specified)
+        """
+        with rest_utils.skip_nested_marshalling():
+            return super(BlueprintsId, self).put(blueprint_id=blueprint_id,
+                                                 **kwargs)
+
+    @swagger.operation(
+        responseClass=models.Blueprint,
+        nickname="deleteById",
+        notes="deletes a blueprint by its id."
+    )
+    @rest_decorators.exceptions_handled
+    @rest_decorators.marshal_with(models.Blueprint)
+    def delete(self, blueprint_id, **kwargs):
+        """
+        Delete blueprint by id
+        """
+        with rest_utils.skip_nested_marshalling():
+            return super(BlueprintsId, self).delete(
+                blueprint_id=blueprint_id, **kwargs)
