+++ rest-service/setup.py
@@ -55,12 +55,15 @@ setup(
     version='4.0a12',
     author='Gigaspaces',
     author_email='cosmo-admin@gigaspaces.com',
-    packages=['manager_rest',
-              'manager_rest.rest',
-              'manager_rest.rest.resources_v1',
-              'manager_rest.deployment_update',
-              'manager_rest.storage',
-              'manager_rest.security'],
+    packages=[
+        'manager_rest',
+        'manager_rest.rest',
+        'manager_rest.rest.resources_v1',
+        'manager_rest.rest.resources_v2',
+        'manager_rest.deployment_update',
+        'manager_rest.storage',
+        'manager_rest.security',
+    ],
     package_data={'manager_rest': ['VERSION']},
     license='LICENSE',
     description='Cloudify manager rest service',
